refactor(ui): the transfer dialog
