+++ packages/maskbook/src/extension/options-page/DashboardComponents/ERC20TokenActionsBar.tsx
@@ -3,7 +3,7 @@ import MoreHorizIcon from '@material-ui/icons/MoreHoriz'
 import { isETH } from '../../../web3/helpers'
 import { useStylesExtends } from '../../../components/custom-ui-helper'
 import { useModal } from '../DashboardDialogs/Base'
-import { DashboardWalletHideTokenConfirmDialog } from '../DashboardDialogs/Wallet'
+import { DashboardWalletHideTokenConfirmDialog, DashboardWalletTransferDialogFT } from '../DashboardDialogs/Wallet'
 import { useMenu } from '../../../utils/hooks/useMenu'
 import type { WalletRecord } from '../../../plugins/Wallet/database/types'
 import { useI18N } from '../../../utils/i18n-next-ui'
@@ -11,7 +11,6 @@ import type { ERC20TokenDetailed, EtherTokenDetailed } from '../../../web3/types
 import { useRemoteControlledDialog } from '../../../utils/hooks/useRemoteControlledDialog'
 import { PluginTransakMessages } from '../../../plugins/Transak/messages'
 import { useAccount } from '../../../web3/hooks/useAccount'
-import { DashboardWalletTransferDialog } from './TransferDialog'
 import { useChainIdValid } from '../../../web3/hooks/useChainState'
 
 const useStyles = makeStyles((theme) => ({
@@ -21,9 +20,9 @@ const useStyles = makeStyles((theme) => ({
 }))
 
 export interface ERC20TokenActionsBarProps extends withClasses<KeysInferFromUseStyles<typeof useStyles>> {
-    wallet: WalletRecord
     chain: 'eth' | string
-    token: ERC20TokenDetailed | EtherTokenDetailed
+    wallet: WalletRecord
+    token: EtherTokenDetailed | ERC20TokenDetailed
 }
 
 export function ERC20TokenActionsBar(props: ERC20TokenActionsBarProps) {
@@ -39,7 +38,7 @@ export function ERC20TokenActionsBar(props: ERC20TokenActionsBarProps) {
     const [, setBuyDialogOpen] = useRemoteControlledDialog(PluginTransakMessages.events.buyTokenDialogUpdated)
     //#endregion
 
-    const [transeferDialog, , openTransferDialogOpen] = useModal(DashboardWalletTransferDialog)
+    const [transeferDialog, , openTransferDialogOpen] = useModal(DashboardWalletTransferDialogFT)
     const [hideTokenConfirmDialog, , openHideTokenConfirmDialog] = useModal(DashboardWalletHideTokenConfirmDialog)
     const [menu, openMenu] = useMenu(
         ...[
