+++ packages/maskbook/src/extension/options-page/DashboardComponents/ERC721TokenActionsBar.tsx
@@ -1,0 +1,63 @@
+import { useCallback } from 'react'
+import { IconButton, makeStyles, MenuItem } from '@material-ui/core'
+import MoreHorizIcon from '@material-ui/icons/MoreHoriz'
+import { useStylesExtends } from '../../../components/custom-ui-helper'
+import { useModal } from '../DashboardDialogs/Base'
+import { DashboardWalletHideTokenConfirmDialog, DashboardWalletTransferDialogNFT } from '../DashboardDialogs/Wallet'
+import { useMenu } from '../../../utils/hooks/useMenu'
+import type { WalletRecord } from '../../../plugins/Wallet/database/types'
+import { useI18N } from '../../../utils/i18n-next-ui'
+import type { ERC721TokenDetailed } from '../../../web3/types'
+import { useAccount } from '../../../web3/hooks/useAccount'
+import { useChainIdValid } from '../../../web3/hooks/useChainState'
+
+const useStyles = makeStyles((theme) => ({
+    more: {
+        color: theme.palette.text.primary,
+    },
+}))
+
+export interface ERC721TokenActionsBarProps extends withClasses<KeysInferFromUseStyles<typeof useStyles>> {
+    wallet: WalletRecord
+    token: ERC721TokenDetailed
+}
+
+export function ERC721TokenActionsBar(props: ERC721TokenActionsBarProps) {
+    const { wallet, token } = props
+
+    const { t } = useI18N()
+    const account = useAccount()
+    const classes = useStylesExtends(useStyles(), props)
+
+    const chainIdValid = useChainIdValid()
+
+    const [transeferDialog, , openTransferDialogOpen] = useModal(DashboardWalletTransferDialogNFT)
+    const [hideTokenConfirmDialog, , openHideTokenConfirmDialog] = useModal(DashboardWalletHideTokenConfirmDialog)
+    const [menu, openMenu] = useMenu(
+        ...[
+            <MenuItem disabled={!chainIdValid} onClick={() => openTransferDialogOpen({ wallet, token })}>
+                {t('transfer')}
+            </MenuItem>,
+            <MenuItem onClick={() => openHideTokenConfirmDialog({ wallet, token })}>{t('hide')}</MenuItem>,
+        ],
+    )
+
+    const onClickButton = useCallback(
+        (ev: React.MouseEvent<HTMLButtonElement>) => {
+            ev.preventDefault()
+            openMenu(ev)
+        },
+        [openMenu],
+    )
+
+    return (
+        <>
+            <IconButton className={classes.more} size="small" onClick={onClickButton}>
+                <MoreHorizIcon />
+            </IconButton>
+            {menu}
+            {hideTokenConfirmDialog}
+            {transeferDialog}
+        </>
+    )
+}
