+++ packages/maskbook/src/plugins/Wallet/database/Wallet.db.ts
@@ -1,23 +1,20 @@
 import { DBSchema, openDB } from 'idb/with-async-ittr-cjs'
-import { createDBAccess, createTransaction } from '../../../database/helpers/openDB'
-import type { ERC20TokenRecord, ERC721TokenRecord, WalletRecordInDatabase } from './types'
+import { createDBAccess } from '../../../database/helpers/openDB'
+import type { ERC20TokenRecordInDatabase, ERC721TokenRecordInDatabase, WalletRecordInDatabase } from './types'
 import type { RedPacketRecordInDatabase } from '../../RedPacket/types'
 import { sideEffect } from '../../../utils/side-effects'
 import { migratePluginDatabase } from './migrate.plugins'
-import { formatChecksumAddress } from '../formatter'
-import { ProviderType } from '../../../web3/types'
 
 function path<T>(x: T) {
     return x
 }
 export const createWalletDBAccess = createDBAccess(() => {
-    return openDB<WalletDB>('maskbook-plugin-wallet', 5, {
+    return openDB<WalletDB>('maskbook-plugin-wallet', 6, {
         async upgrade(db, oldVersion, newVersion, tx) {
             function v0_v1() {
                 // @ts-expect-error
                 db.createObjectStore('RedPacket', { keyPath: path<keyof RedPacketRecordInDatabase>('id') })
-                db.createObjectStore('ERC20Token', { keyPath: path<keyof WalletRecordInDatabase>('address') })
-                db.createObjectStore('ERC721Token', { keyPath: path<keyof WalletRecordInDatabase>('address') })
+                db.createObjectStore('ERC20Token', { keyPath: path<keyof ERC20TokenRecordInDatabase>('address') })
                 db.createObjectStore('Wallet', { keyPath: path<keyof WalletRecordInDatabase>('address') })
             }
             function v1_v2() {
@@ -35,7 +32,7 @@ export const createWalletDBAccess = createDBAccess(() => {
              *     }
              * }
              */
-            async function v2_v3() {
+            function v2_v3() {
                 const os = db.createObjectStore('PluginStore', { keyPath: 'record_id' })
                 // @ts-ignore
                 os.createIndex('0', 'index0')
@@ -49,50 +46,18 @@ export const createWalletDBAccess = createDBAccess(() => {
                 // @ts-ignore
                 db.deleteObjectStore('RedPacket')
             }
-            /**
-             * Use checksummed address in DB
-             */
-            async function v3_v4() {
-                const t = createTransaction(db, 'readwrite')('Wallet', 'ERC20Token')
-                const wallets = t.objectStore('Wallet')
-                const tokens = t.objectStore('ERC20Token')
-                for await (const wallet of wallets) {
-                    // update address
-                    wallet.value.address = formatChecksumAddress(wallet.value.address)
 
-                    // update token list sets
-                    ;[wallet.value.erc20_token_blacklist, wallet.value.erc20_token_whitelist].forEach((set) => {
-                        const values = Array.from(set.values())
-                        set.clear()
-                        values.forEach((value) => set.add(formatChecksumAddress(value)))
-                    })
-                    await wallet.update(wallet.value)
-                }
-                for await (const token of tokens) {
-                    token.value.address = formatChecksumAddress(token.value.address)
-                    await token.update(token.value)
-                }
-            }
             /**
-             * Fix providerType does not exist in legacy wallet
+             * Store ERC721Token records in DB
              */
-            async function v4_v5() {
-                const t = createTransaction(db, 'readwrite')('Wallet')
-                const wallets = t.objectStore('Wallet')
-                for await (const wallet of wallets) {
-                    const wallet_ = wallet as any
-                    if (wallet_.value.provider) continue
-                    if (wallet_.value.type === 'managed') wallet_.value.provider = ProviderType.Maskbook
-                    else if (wallet_.value.type === 'exotic') wallet_.value.provider = ProviderType.MetaMask
-                    await wallet.update(wallet_.value)
-                }
+            function v5_v6() {
+                db.createObjectStore('ERC721Token', { keyPath: path<keyof ERC721TokenRecordInDatabase>('record_id') })
             }
 
             if (oldVersion < 1) v0_v1()
             if (oldVersion < 2) v1_v2()
-            if (oldVersion < 3) await v2_v3()
-            if (oldVersion < 4) await v3_v4()
-            if (oldVersion < 5) await v4_v5()
+            if (oldVersion < 3) v2_v3()
+            if (oldVersion < 6) v5_v6()
         },
     })
 })
@@ -118,11 +83,11 @@ export interface WalletDB<Data = unknown, Indexes extends [IDBValidKey?, IDBVali
         key: string
     }
     ERC20Token: {
-        value: ERC20TokenRecord
+        value: ERC20TokenRecordInDatabase
         key: string
     }
     ERC721Token: {
-        value: ERC721TokenRecord
+        value: ERC721TokenRecordInDatabase
         key: string
     }
 }
