+++ packages/maskbook/src/plugins/Wallet/hooks/useERC721Token.ts
@@ -1,30 +1,30 @@
 import { ValueRef } from '@dimensiondev/holoflows-kit'
 import { WalletMessages, WalletRPC } from '../messages'
 import { useValueRef } from '../../../utils/hooks/useValueRef'
-import type { ERC20TokenRecord } from '../database/types'
-import { ERC20TokenArrayComparer } from '../helpers'
+import type { ERC721TokenRecord } from '../database/types'
+import { ERC721TokenArrayComparer } from '../helpers'
 import { useWallet } from './useWallet'
-import { ERC20TokenDetailed, EthereumTokenType } from '../../../web3/types'
+import { ERC721TokenDetailed, EthereumTokenType } from '../../../web3/types'
 
 //#region cache service query result
-const erc20TokensRef = new ValueRef<ERC20TokenRecord[]>([], ERC20TokenArrayComparer)
+const erc721TokensRef = new ValueRef<ERC721TokenRecord[]>([], ERC721TokenArrayComparer)
 
 async function revalidate() {
-    // erc20 tokens
-    const erc20Tokens = await WalletRPC.getERC20Tokens()
-    erc20TokensRef.value = erc20Tokens
+    // erc721 tokens
+    const erc721Tokens = await WalletRPC.getERC721Tokens()
+    erc721TokensRef.value = erc721Tokens
 }
 WalletMessages.events.erc20TokensUpdated.on(revalidate)
 revalidate()
 //#endregion
 
 /**
- * Fetch all ERC20 tokens from DB
+ * Fetch all ERC721 tokens from DB
  */
-export function useERC20TokensFromDB(): ERC20TokenDetailed[] {
-    const records = useValueRef(erc20TokensRef)
+export function useERC721TokensFromDB(): ERC721TokenDetailed[] {
+    const records = useValueRef(erc721TokensRef)
     return records.map((x) => ({
-        type: EthereumTokenType.ERC20,
+        type: EthereumTokenType.ERC721,
         ...x,
     }))
 }
@@ -33,9 +33,9 @@ export function useERC20TokensFromDB(): ERC20TokenDetailed[] {
  * Fetch all trusted ERC20 tokens from DB
  * @param address
  */
-export function useTrustedERC20TokensFromDB() {
+export function useTrustedERC721TokensFromDB() {
     const wallet = useWallet()
-    const tokens = useERC20TokensFromDB()
+    const tokens = useERC721TokensFromDB()
 
     if (!wallet) return []
     return tokens.filter(
