+++ packages/maskbook/src/plugins/Wallet/services/erc721.ts
@@ -20,27 +20,28 @@ export async function addERC721Token(token: ERC721TokenDetailed) {
             symbol: token.symbol ?? '',
         }),
     )
-    WalletMessages.events.tokensUpdated.sendToAll(undefined)
+    WalletMessages.events.erc721TokensUpdated.sendToAll(undefined)
 }
 
 export async function removeERC721Token(token: PartialRequired<ERC721TokenDetailed, 'address'>) {
     const t = createTransaction(await createWalletDBAccess(), 'readwrite')('ERC721Token', 'Wallet')
     await t.objectStore('ERC721Token').delete(formatChecksumAddress(token.address))
-    WalletMessages.events.tokensUpdated.sendToAll(undefined)
+    WalletMessages.events.erc721TokensUpdated.sendToAll(undefined)
 }
 
 export async function trustERC721Token(address: string, token: ERC721TokenDetailed) {
     const t = createTransaction(await createWalletDBAccess(), 'readwrite')('ERC721Token', 'Wallet')
     const wallet = await getWalletByAddress(t, formatChecksumAddress(address))
     assert(wallet)
-    const tokenAddressChecksummed = formatChecksumAddress(token.address)
     let updated = false
-    if (!wallet.erc721_token_whitelist.has(tokenAddressChecksummed)) {
-        wallet.erc721_token_whitelist.add(tokenAddressChecksummed)
+    const tokenAddressChecksummed = formatChecksumAddress(token.address)
+    const key = `${tokenAddressChecksummed}_${token.tokenId}`
+    if (!wallet.erc721_token_whitelist.has(key)) {
+        wallet.erc721_token_whitelist.add(key)
         updated = true
     }
-    if (wallet.erc721_token_blacklist.has(tokenAddressChecksummed)) {
-        wallet.erc721_token_blacklist.delete(tokenAddressChecksummed)
+    if (wallet.erc721_token_blacklist.has(key)) {
+        wallet.erc721_token_blacklist.delete(key)
         updated = true
     }
     if (!updated) return
@@ -48,18 +49,22 @@ export async function trustERC721Token(address: string, token: ERC721TokenDetail
     WalletMessages.events.walletsUpdated.sendToAll(undefined)
 }
 
-export async function blockERC721Token(address: string, token: PartialRequired<ERC721TokenDetailed, 'address'>) {
+export async function blockERC721Token(
+    address: string,
+    token: PartialRequired<ERC721TokenDetailed, 'address' | 'tokenId'>,
+) {
     const t = createTransaction(await createWalletDBAccess(), 'readwrite')('ERC721Token', 'Wallet')
     const wallet = await getWalletByAddress(t, formatChecksumAddress(address))
     assert(wallet)
     let updated = false
     const tokenAddressChecksummed = formatChecksumAddress(token.address)
-    if (wallet.erc721_token_whitelist.has(tokenAddressChecksummed)) {
-        wallet.erc721_token_whitelist.delete(tokenAddressChecksummed)
+    const key = `${tokenAddressChecksummed}_${token.tokenId}`
+    if (wallet.erc721_token_whitelist.has(key)) {
+        wallet.erc721_token_whitelist.delete(key)
         updated = true
     }
-    if (!wallet.erc721_token_blacklist.has(tokenAddressChecksummed)) {
-        wallet.erc721_token_blacklist.add(tokenAddressChecksummed)
+    if (!wallet.erc721_token_blacklist.has(key)) {
+        wallet.erc721_token_blacklist.add(key)
         updated = true
     }
     if (!updated) return
