+++ packages/maskbook/src/web3/helpers.ts
@@ -10,6 +10,7 @@ import {
     EtherTokenDetailed,
     CurrencyType,
     AssetDetailed,
+    ERC721TokenDetailed,
 } from './types'
 
 export function isSameAddress(addrA: string, addrB: string) {
@@ -78,6 +79,27 @@ export function createERC20Token(chainId: ChainId, address: string, decimals: nu
         name,
         symbol,
     } as ERC20TokenDetailed
+}
+
+export function createERC721Token(
+    chainId: ChainId,
+    tokenId: string,
+    address: string,
+    name: string,
+    symbol: string,
+    baseURI?: string,
+    tokenURI?: string,
+) {
+    return {
+        type: EthereumTokenType.ERC721,
+        chainId,
+        tokenId,
+        address,
+        name,
+        symbol,
+        baseURI,
+        tokenURI,
+    } as ERC721TokenDetailed
 }
 
 export function decodeOutputString(web3: Web3, abis: AbiOutput[], output: string) {
